@@ -8,11 +8,11 @@ apply plugin: 'com.google.gms.google-services'
 
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.joshtalks.joshskills"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 20
         versionName "1.0.20"
         vectorDrawables.useSupportLibrary true
@@ -97,9 +97,8 @@ android {
         exclude 'META-INF/ASL2.0'
         exclude 'META-INF/notice.txt'
         exclude 'META-INF/main.kotlin_module'
-
-
     }
+
     lintOptions {
         abortOnError false
     }
@@ -202,7 +201,7 @@ dependencies {
 
 
     implementation 'com.facebook.shimmer:shimmer:0.4.0'
-    implementation 'com.facebook.android:account-kit-sdk:5.+'
+    implementation 'com.facebook.android:account-kit-sdk:5.4.0'
 
 
     implementation 'com.facebook.android:facebook-android-sdk:5.12.0'
@@ -233,7 +232,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.6.1'
     implementation 'com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:0.9.2'
     implementation 'com.github.esafirm.android-image-picker:imagepicker:2.1.0'
-    implementation("com.squareup.okhttp3:logging-interceptor:4.1.0")
+    implementation 'com.squareup.okhttp3:logging-interceptor:4.1.0'
 
 
     implementation 'com.mindorks.android:placeholderview:1.0.3'
