@@ -6,6 +6,7 @@ import android.graphics.drawable.ShapeDrawable
 import android.graphics.drawable.shapes.OvalShape
 import android.graphics.drawable.shapes.RectShape
 import android.graphics.drawable.shapes.RoundRectShape
+import java.util.*
 
 class TextDrawable private constructor(builder: Builder) : ShapeDrawable(builder.shape) {
 
@@ -29,7 +30,7 @@ class TextDrawable private constructor(builder: Builder) : ShapeDrawable(builder
         radius = builder.radius
 
         // text and color
-        text = if (builder.toUpperCase) builder.text!!.toUpperCase() else builder.text
+        text = if (builder.toUpperCase) builder.text!!.toUpperCase(Locale.getDefault()) else builder.text
         color = builder.color
 
         // text paint settings
