@@ -26,7 +26,7 @@ public interface MediaFormatStrategy {
      * @return null for passthrough.
      * @throws OutputFormatUnavailableException if input could not be transcoded because of restrictions.
      */
-    public MediaFormat createVideoOutputFormat(MediaFormat inputFormat);
+    MediaFormat createVideoOutputFormat(MediaFormat inputFormat);
 
     /**
      * Caution: this method should return null currently.
@@ -34,6 +34,6 @@ public interface MediaFormatStrategy {
      * @return null for passthrough.
      * @throws OutputFormatUnavailableException if input could not be transcoded because of restrictions.
      */
-    public MediaFormat createAudioOutputFormat(MediaFormat inputFormat);
+    MediaFormat createAudioOutputFormat(MediaFormat inputFormat);
 
 }
